Fix issue with exponential backoff returning floating value in Python 2

+++ sqs_workers/backoff_policies.py
@@ -38,7 +38,7 @@ class ExponentialBackoff(object):
                                  visibility_timeout)
         visibility_timeout = min(self.max_visibility_timeout,
                                  visibility_timeout)
-        return round(visibility_timeout)
+        return int(round(visibility_timeout))
 
 
 DEFAULT_BACKOFF = ExponentialBackoff()
